Rework of fiscal sponsorship page (#1238)
